refactor party modal and KamiCard
