+++ packages/client/src/layers/react/components/library/KamiCard.tsx
@@ -1,7 +1,10 @@
-import React from 'react';
-import styled, { keyframes } from 'styled-components';
+import React, { useEffect, useState } from 'react';
+import styled from 'styled-components';
 
-import clickSoundUrl from 'assets/sound/fx/mouseclick.wav';
+import { Battery } from './Battery';
+import { Countdown } from './Countdown';
+import { Tooltip } from './Tooltip';
+import { Card } from 'layers/react/components/library/Card';
 import { Kami } from 'layers/react/shapes/Kami';
 import { dataStore } from 'layers/react/store/createStore';
 
@@ -11,6 +14,8 @@ interface Props {
   subtext?: string;
   action?: React.ReactNode;
   cornerContent?: React.ReactNode;
+  battery?: boolean;
+  cooldown?: boolean;
 }
 
 // KamiCard is a card that displays information about a Kami. It is designed to display
@@ -21,15 +26,24 @@ export const KamiCard = (props: Props) => {
     setVisibleModals,
     selectedEntities,
     setSelectedEntities,
-    sound: { volume },
   } = dataStore();
+  const [lastRefresh, setLastRefresh] = useState(Date.now());
 
-  // layer on a sound effect
-  const playClickAudio = async () => {
-    const clickSound = new Audio(clickSoundUrl);
-    clickSound.volume = volume * 0.6;
-    clickSound.play();
-  };
+
+  // ticking
+  useEffect(() => {
+    const refreshClock = () => {
+      setLastRefresh(Date.now());
+    };
+    const timerId = setInterval(refreshClock, 1000);
+    return function cleanup() {
+      clearInterval(timerId);
+    };
+  }, []);
+
+
+  /////////////////
+  // INTERACTION
 
   // toggle the kami modal settings depending on current its current state
   const kamiOnClick = () => {
@@ -45,87 +59,86 @@ export const KamiCard = (props: Props) => {
       setSelectedEntities({ ...selectedEntities, kami: props.kami.entityIndex });
       setVisibleModals({ ...visibleModals, kami: true });
     }
-    playClickAudio();
   }
+
+  /////////////////
+  // INTERPRETATION
+
+  // calculate the time a kami has spent idle (in seconds)
+  const calcIdleTime = (kami: Kami): number => {
+    return lastRefresh / 1000 - kami.lastUpdated;
+  };
+
+  // calculate health based on the drain against last confirmed health
+  const calcHealth = (kami: Kami): number => {
+    let health = 1 * kami.health;
+    let duration = calcIdleTime(kami);
+    health += kami.healthRate * duration;
+    health = Math.min(Math.max(health, 0), kami.stats.health);
+    return health;
+  };
+
+  // check whether the kami is revealed
+  const isUnrevealed = (kami: Kami): boolean => {
+    return kami.state === 'UNREVEALED';
+  };
+
+
+  /////////////////
+  // DISPLAY
 
   // generate the styled text divs for the description
   const Description = () => {
     const header = [<TextBig key='header'>{props.description[0]}</TextBig>];
-
     const details = props.description
       .slice(1)
       .map((line, index) => <TextMedium key={`description-${index}`}>{line}</TextMedium>);
     return [...header, ...details];
   };
 
+  const CornerContent = (kami: Kami) => {
+    const healthString = !isUnrevealed(kami)
+      ? `Health: ${calcHealth(kami).toFixed()}/${kami.stats.health * 1}`
+      : '';
+
+    const cooldown = Math.round(Math.max(kami.cooldown - calcIdleTime(kami), 0));
+    const cooldownString = `Cooldown: ${Math.max(cooldown, 0).toFixed(0)}s`;
+    return (
+      <TitleCorner>
+        {props.cooldown &&
+          <Tooltip text={[cooldownString]}>
+            <Countdown total={kami.cooldown} current={cooldown} />
+          </Tooltip>
+        }
+        {props.battery &&
+          <Tooltip text={[healthString]}>
+            <Battery level={100 * calcHealth(kami) / kami.stats.health} />
+          </Tooltip>
+        }
+      </TitleCorner>
+    );
+  };
+
   return (
-    <Card key={props.kami.id}>
-      <Image onClick={() => kamiOnClick()} src={props.kami.uri} />
-      <Container>
-        <TitleBar>
-          <TitleText onClick={() => kamiOnClick()}>{props.kami.name}</TitleText>
-          <TitleCorner>{props.cornerContent}</TitleCorner>
-        </TitleBar>
-        <Content>
-          <ContentColumn>{Description()}</ContentColumn>
-          <ContentColumn>
-            <ContentSubtext>{props.subtext}</ContentSubtext>
-            <ContentActions>{props.action}</ContentActions>
-          </ContentColumn>
-        </Content>
-      </Container>
-    </Card>
+    <Card
+      image={props.kami.uri}
+      imageOnClick={() => kamiOnClick()}
+      titleBarContent={[
+        <TitleText onClick={() => kamiOnClick()}>{props.kami.name}</TitleText>,
+        CornerContent(props.kami)
+      ]}
+      content={[
+        <ContentColumn key='column-1'>{Description()}</ContentColumn>,
+        <ContentColumn key='column-2'>
+          <ContentSubtext>{props.subtext}</ContentSubtext>
+          <ContentActions>{props.action}</ContentActions>
+        </ContentColumn>
+      ]}
+    />
   );
 };
 
-const Card = styled.div`
-  background-color: #fff;
-  border-color: black;
-  border-radius: .35vw;
-  border-style: solid;
-  border-width: .15vw;
-  color: black;
-  margin: .3vw .15vw 0vw .15vw;
-
-  display: flex;
-  flex-flow: row nowrap;
-`;
-
-const Image = styled.img`
-  border-style: solid;
-  border-width: 0vw .15vw 0vw 0vw;
-  border-color: black;
-  border-radius: .15vw 0vw 0vw .15vw;
-  height: 9vw;
-
-  &:hover {
-    opacity: 0.75;
-  }
-`;
-
-const Container = styled.div`
-  border-color: black;
-  border-width: .15vw;
-  color: black;
-  flex-grow: 1;
-
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: stretch;
-`;
-
-const TitleBar = styled.div`
-  border-style: solid;
-  border-width: 0vw 0vw .15vw 0vw;
-  border-color: black;
-  padding: .45vw;
-
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: center;
-`;
-
-const TitleText = styled.p`
+const TitleText = styled.div`
   font-family: Pixel;
   font-size: 1vw;
   text-align: left;
@@ -147,15 +160,6 @@ const TitleCorner = styled.div`
   display: flex;
   align-items: center;
   justify-content: flex-end;
-`;
-
-const Content = styled.div`
-  flex-grow: 1;
-  padding: .7vw;
-
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: stretch;
 `;
 
 const ContentColumn = styled.div`
@@ -193,29 +197,4 @@ const TextMedium = styled.p`
   text-align: left;
   padding-top: .4vw;
   padding-left: .2vw;
-`;
-
-// lol
-const rotate = keyframes`
-  from {
-    transform: rotate(0deg);
-  }
-
-  to {
-    transform: rotate(360deg);
-  }
-`;
-
-const ImageBackflip = styled.img`
-  border-style: solid;
-  border-width: 0px 2px 0px 0px;
-  border-color: black;
-  height: 110px;
-  margin: 0px;
-  padding: 0px;
-
-  &:click {
-    animation: ${rotate} 0.3s linear infinite;
-    animation-iteration-count: 1;
-  }
 `;
