+++ packages/client/src/layers/react/components/modals/party/index.ts
@@ -1,0 +1,3 @@
+import { registerPartyModal } from "./Party";
+
+export { registerPartyModal } from "./Party";
